+++ v_4_9_0/render_asset_content_test.go
@@ -1,4 +1,4 @@
-package v_4_8_0
+package v_4_9_0
 
 import (
 	"strings"
