+++ v_4_9_0/worker_template.go
@@ -1,4 +1,4 @@
-package v_4_8_0
+package v_4_9_0
 
 const WorkerTemplate = `---
 ignition:
